requirements.txt
streamlit==1.32.0
pandas==2.2.0
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
google-generativeai==0.3.1
python-dotenv==1.0.0
plotly==5.18.0